% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adder.R
\name{adder}
\alias{adder}
\title{Add two numbers}
\usage{
adder(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number!}
}
\value{
Returns the result of addition
}
\description{
This is a riveting description that describes what your function does. This
is a link to \code{\link{hello}}.
}
\examples{
adder(1, 2)
}
\seealso{
\code{\link{hello}}
}

